+++ AdventureBackpacks/Patches/Humanoid.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Reflection.Emit;
 using System.Threading;
+using AdventureBackpacks.Assets;
 using AdventureBackpacks.Components;
 using AdventureBackpacks.Extensions;
 using AdventureBackpacks.Features;
@@ -134,13 +135,16 @@ public class HumanoidPatches
             var player = Player.m_localPlayer;
             var item = __0;
 
-            if (item.IsBackpack())
+            if (item.IsBackpack() && item.TryGetBackpackItem(out var backpack))
             {
                 InventoryGuiPatches.BackpackEquipped = true;
                 
-                var backpackContainer = player.gameObject.GetComponent<Container>();
-                var backpack = item.Data().GetOrCreate<BackpackComponent>();
-                backpack.UpdateContainerSizing(backpackContainer);
+                var backpackItem = item.Data().GetOrCreate<BackpackComponent>();
+                
+                if (!Backpacks.IsEmptyingBackpack && backpackItem.InventoryNeedsValidating(backpack.BackpackSize[backpackItem.Item.m_quality].Value))
+                {
+                    Backpacks.ValidateBackpackInventorySizing(player, backpackItem.Item);
+                }
             }
         }
     }